Moved landing page components to home page
